requirements.txt
feedparser
requests
pyyaml
beautifulsoup4
python-dateutil
